pyproject.toml
[project]
name = "end-to-end-agentic-ai-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
dev = [
    "end-to-end-agentic-ai-project",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-google-community[places]>=2.0.7",
    "langchain-groq>=0.3.6",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.28",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.2",
    "langsmith>=0.4.10",
    "o365>=2.1.4",
    "openai>=1.98.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "pymongo>=4.13.2",
    "pymupdf>=1.26.3",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.47.1",
    "uvicorn>=0.35.0",
]

[tool.uv.sources]
end-to-end-agentic-ai-project = { workspace = true }
